telegram-mantra-bot/requirements.txt
aiogram==3.0.0
openai==0.27.0
SQLAlchemy==2.0.15
#psycopg2-binary==2.9.6
python-dotenv==1.0.0
alembic==1.13.0
